otel instrumentation more centralized location, use in both main.sh files
