+++ PLATER/plater-clustered-docker/Dockerfile
@@ -4,7 +4,6 @@ USER root
 ENV MODE="deploy"
 RUN pip install gunicorn
 COPY ./main.sh /home/plater/Plater/
-COPY ./gunicorn.conf.py /home/plater/Plater/
 RUN chmod 777 /home/plater/Plater/main.sh
 COPY ./heartbeat.py /home/plater/Plater/PLATER/services/
 USER nru
