+++ PLATER/services/otel.py
@@ -1,0 +1,46 @@
+"""OpenTelemetry tracing setup for the Plater FastAPI app.
+
+Must be called *after* the gunicorn fork (from a post_fork hook) so the
+BatchSpanProcessor background thread and gRPC channel are created in the
+worker process, not the master. See
+https://oneuptime.com/blog/post/2026-02-06-troubleshoot-fastapi-uvicorn-reload/view
+"""
+
+
+def setup_otel(worker_pid=None):
+    from PLATER.services.config import config
+
+    if config.get("OTEL_ENABLED", "False") in ("false", "False"):
+        return
+
+    from opentelemetry import trace
+    from opentelemetry.sdk.resources import SERVICE_NAME, Resource
+    from opentelemetry.sdk.trace import TracerProvider
+    from opentelemetry.sdk.trace.export import BatchSpanProcessor
+    from opentelemetry.instrumentation.fastapi import FastAPIInstrumentor
+
+    PLATER_TITLE = config.get('PLATER_TITLE', 'Plater API')
+    resource_attrs = {SERVICE_NAME: config.get("OTEL_SERVICE_NAME", PLATER_TITLE)}
+    if worker_pid is not None:
+        resource_attrs["worker.pid"] = worker_pid
+    resource = Resource.create(attributes=resource_attrs)
+    provider = TracerProvider(resource=resource)
+
+    OTEL_USE_CONSOLE_EXPORTER = config.get(
+        "OTEL_USE_CONSOLE_EXPORTER", "False"
+    ) not in ("false", "False")
+
+    if OTEL_USE_CONSOLE_EXPORTER:
+        from opentelemetry.sdk.trace.export import ConsoleSpanExporter
+        processor = BatchSpanProcessor(ConsoleSpanExporter())
+    else:
+        from opentelemetry.exporter.otlp.proto.grpc.trace_exporter import OTLPSpanExporter
+        otlp_host = config.get("JAEGER_HOST", "http://localhost").rstrip('/')
+        otlp_port = config.get("JAEGER_PORT", "4317")
+        processor = BatchSpanProcessor(OTLPSpanExporter(endpoint=f'{otlp_host}:{otlp_port}'))
+
+    provider.add_span_processor(processor)
+    trace.set_tracer_provider(provider)
+
+    from PLATER.services.app_trapi import APP
+    FastAPIInstrumentor.instrument_app(APP, excluded_urls="docs,openapi.json")
